requirements.txt
gradio==3.50.2
openai==1.3.5
anthropic==0.8.1
google-generativeai==0.3.1
transformers==4.35.2
torch==2.1.1
requests==2.31.0
pydub==0.25.1
numpy==1.26.2
Pillow==10.1.0